requirements.txt
Django==1.8.4
django-mptt==0.7.4
djangorestframework==3.3.0
Pillow==3.0.0
psycopg2==2.6.1
wheel==0.26.0
